services/vpngw/src/pyproject.toml
[tool.poetry]
name = "nebius-vpngw"
version = "0.1.0"
description = "Nebius VM-based VPN Gateway orchestrator and agent"
authors = ["Nebius <dev@nebius.com>"]
readme = "README.md"
packages = [{ include = "nebius_vpngw", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
PyYAML = "^6.0.1"
paramiko = "^3.5.0"
rich = "^13.9.2"
typer = {version = "^0.12.5", extras=["all"]}
pydantic = "^2.9.2"

# Nebius SDK placeholder; replace with actual package name when available
nebius-pysdk = "^0.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.6.9"

[tool.poetry.scripts]
nebius-vpngw = "nebius_vpngw.cli:app"
nebius-vpngw-agent = "nebius_vpngw.agent.main:main"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"